Simplify sample rate (iPhone 6s) fix and fix writing to the camera roll.
When you set overrideOutputAudioPort on the shared AVAudioSession to AVAudioSessionPortOverrideNone, the sample rate is set to the expected 44100. Since that's the case, there's no need to tinker with the buffer duration settings.

The camera roll writing has been failing (since iOS 9 I guess) with -11800 and -12902 error. Removing the unnecessary compression settings for its asset writer seems to have fixed that bug.